+++ data-raw/create_internal_constants.R
@@ -1,0 +1,6 @@
+lims_prop <- c(0, 1)
+lims_perc <- c(0, 100)
+
+usethis::use_data(lims_prop, 
+                  lims_perc,
+                  internal = TRUE, overwrite = TRUE)